requirements.txt
alembic==1.13.1
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.2.2
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.29
typing_extensions==4.10.0
urllib3==2.2.1

